+++ apps/question-bank/src/main/java/gov/cdc/nbs/questionbank/pagerules/PageRuleService.java
@@ -2,7 +2,9 @@ package gov.cdc.nbs.questionbank.pagerules;
 
 import gov.cdc.nbs.questionbank.model.CreateRuleRequest;
 
+import java.math.BigInteger;
+
 public interface PageRuleService {
 
-   Long createPageRule(CreateRuleRequest.ruleRequest request);
+   BigInteger createPageRule(CreateRuleRequest.ruleRequest request);
 }
